[FIX] l10n_fr_fec_oca: fix wrong data
Fix missing data due to a line of code that was accessing the ORM
between cr.execute() and cr.fetchall(), so cr.fetchall() was always
empty. I'm really sorry for that...

Improve the 120/129 line: auto-switch to 129000 with appropriate label
when it's a loss
Add complete support for the option about partners, including in the
initial balances.