[FIX] vat_return: fully ignore cancel moves
In a recent update of l10n_fr_account_vat_return, I changed to code to
avoid raising an error when cancel moves where present. But, in the code
that computed vat on payment, there was no filter on the state on
account.move and account.move.line, so it was taking into account cancel
moves. This commit fixes this.

Improve a variable name.