[IMP] vat_return: warn if draft moves are present (instead of raising an error)
When draft moves where present before end_date of the VAT return, we
used to raise an error. With this commit, we continue to raise an error if option
'fr_vat_update_lock_dates' is enabled on the accounting config page or if one of the draft
moves is the move of the previous VAT return. Otherwise, we start a
wizard that shows the draft moves and gives the choice to the user between
processing the draft moves or ignoring them.