[IMP] account_banking_fr_lcr: add support for paper letter of exchange (accepted letter of exchange)
Add button on customer invoices with payment mode code = 'fr_lcr' and fr_lcr_type = 'accepted' to get the paper letter of change as PDF.
Add support for Dailly convention and cash discounts with value date on payment order (with default values on payment mode).
Add field on customer invoice to explicitely select the bank account of the customer that will be debited.
Improve CFONB file generation to be more compliant with the specs, in particular on references.
Add SIREN of company and customer in CFONB file if l10n_fr_siret module is installed and the SIREN field is set on the partner.
Update FR translation
Update README